moved theme to helpers
